requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.7.4
pydantic-settings==2.4.0
python-multipart==0.0.6

# Frontend
streamlit==1.30.0

# AWS Bedrock
boto3==1.34.34
botocore==1.34.34

# Vector Database and Embeddings
faiss-cpu==1.13.0
numpy==1.26.4

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2
Pillow>=10.4.0
unstructured>=0.15.0
pymupdf>=1.24.0

# Text Processing
tiktoken>=0.8.0
langchain>=0.3.0
langchain-community>=0.3.0

# Logging and Monitoring
loguru==0.7.2
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Testing (optional but recommended)
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Selenium for generated scripts
selenium==4.17.2
webdriver-manager==4.0.1
